+++ R/to_from_all.R
@@ -45,7 +45,7 @@ supported_classes <- function() {
 #' relevant_class(x)
 #' @export
 relevant_class <- function(x) {
-  stopifnot(ts_boxable(x))
+  check_ts_boxable(x)
   intersect(class(x), supported_classes())[1]
 }
 
@@ -55,13 +55,30 @@ relevant_class <- function(x) {
 #' Mainly used internally.
 #'
 #' @inherit ts_default
-#' @return logical, either `TRUE` or `FALSE`
+#' @return logical, either `TRUE` or `FALSE`. `check_ts_boxable()` fails if not
+#'   `TRUE`
 #' @examples
 #' ts_boxable(AirPassengers)
 #' ts_boxable(lm)
 #' @export
 ts_boxable <- function(x) {
   any(supported_classes() %in% class(x))
+}
+
+#' Error Check Functions
+#'
+#' @inherit ts_default
+#'
+#' @name ts_boxable
+#' @export
+check_ts_boxable <- function(x) {
+  if (!ts_boxable(x)) {
+    stop0(
+      "object is of non-ts-boxable class(es) ",
+      paste(paste0("'", class(x), "'"), collapse = ", "),
+      ". See `?ts_ts`."
+    )
+  }
 }
 
 
@@ -115,8 +132,8 @@ desired_class <- function(ll) {
 #'
 #' @inherit ts_default
 #' @param template ts-boxable time series, an object of class `ts`, `xts`,
-#'   `zoo`, `data.frame`, `data.table`, `tbl`, `tbl_ts`, `tbl_time`, `tis`,
-#'   `irts` or `timeSeries`.
+#'   `zoo`, `zooreg`, `data.frame`, `data.table`, `tbl`, `tbl_ts`, `tbl_time`,
+#'   `tis`, `irts` or `timeSeries`.
 #' @param preserve.mode should the mode the time column be preserved
 #'   (data frame only)
 #' @param preserve.names should the name of the time column be preserved