+++ R/ts_first_of_period.R
@@ -31,6 +31,7 @@ dts_first_of_period <- function(x) {
 
   value <- NULL
   has.value <- NULL
+  time.orig <- NULL
   smry <- ts_summary(x)
   start <- date_year(smry$start)
   end <- as.Date(paste(data.table::year(smry$end) + 2, "1", "1", sep = "-"))