+++ R/ts_pick.R
@@ -23,7 +23,7 @@
 #'
 #' @export
 ts_pick <- function(x, ...) {
-  stopifnot(ts_boxable(x))
+  check_ts_boxable(x)
 
   id <- NULL
   call.names <- unlist(lapply(substitute(placeholderFunction(...))[-1], deparse,